Move common support to the system module
Move --host, --build, --prefix and feature support
out of the 'cc' module and into the 'system' module.
The system module also now defines top_srcdir and SHELL
for better autoconf compatibility.

Allow autosetup --help=module to show the help for a
module even if it isn't used in the local auto.def

Signed-off-by: Steve Bennett <steveb@workware.net.au>