cc: drop empty -cflags, -includes, -libs
This addresses the following problem:

I am having a problem with a malformed compiler command line coming from the
fact that I am looping over a set of variable and trying different combinations.
This comes from a module I'm writing which is supposed to locate iconv() either
i) in prefix, ii) in a dedicated libiconv.so library, or iii) in libc.
I am looping over different variations of cflags, ldflags, and libs, any of
which could be empty in a particular iteration. When this happens, the compiler
gets invoked with empty arguments, which end up in it erroring with "No such
file".

Reported-By: Pietro Cerutti <gahr@gahr.ch>
Signed-off-by: Steve Bennett <steveb@workware.net.au>