options: clarify handling of string option defaults
Explicitly document that there are three accepted forms:

  name:value
  name:value=default
  name:=optvalue

And the second form, name:value=default provides a default for the option
even if not specified by the user. Previously this case was not setting
the option value to the default if not specified by the user.

Note that there is no sense in supporting name:=optvalue=default, since specifying
--name is the same as not specifying it at all. Both will resulting in
the option having the default value of 'default'

Update testsuite/testoptions tests to match the above.

Signed-off-by: Steve Bennett <steveb@workware.net.au>