define-append-argv for better handling of args with spaces
Previously autosetup played fast and loose with variables that may end
up on the command line. And that mostly worked, except when values contain
spaces, and then the results could be unexpected.

To address this, add define-append-argv that has clear semantics for
values that contain spaces in a way that the result can be passed to the
shell command line.

For example, consider the previous approach where the same value was
appended more than once.

  define-append ABC "CFLAGS=-Werror -Wall"
  define-append ABC "CFLAGS=-Werror -Wall"

We end up with:

  ABC=CFLAGS=-Werror -Wall CFLAGS=-Werror -Wall

This is due to the way duplicate detection was implemented.
This is especially problematic with the 'cc' module adding CC, CFLAGS, etc.
from the environment/command line to AUTOREMAKE as these values are very
likely to contains spaces.

Now define-append-argv is designed specifically to create variables destined
for the shell command line, with correct duplicate detection and ensuring
that the result is always well quoted.

  define-append-argv ABC "CFLAGS=-Werror -Wall"
  define-append-argv ABC "CFLAGS=-Werror -Wall"

Results in:

  ABC="CFLAGS=-Werror -Wall"

Use this when constructing AUTOREMAKE and CONFIGURE_OPTS

Signed-off-by: Steve Bennett <steveb@workware.net.au>